Made piece spawning after line clears more stable (maybe?)
Sometimes it would spawn 2 pieces in quick succession causing an immdiate topout, to try prevent this I made the line clear animation itself the waiting period, rather than having a separate "await" call for just the piece spawn delay. Not even sure if that was the cause, but since this bug is rare I can only push this change and see if play testing shows any improvement after this update.