Added mouse controls for gameplay
Scroll down to move left, scroll up to move right, middle mouse button to soft drop.
This is more for proof of concept and fun, probably wont be a final feature, and if it is, it will be toggleable since it gives a significant advantage over tapping and even rolling. May consider making high scores not save while its toggled on if I end up leaving it as a togglable option.